server: esajpip - fix build errors uncovered by Darwin 13 compiler (final)
git-svn-id: svn+ssh://solsvn.oma.be/home/sidc/ssr/SWHV_repos/trunk/server/esajpip@310 b4e469a2-07ce-4b26-9273-4d7d95a670c7

+++ log4cpp/log4cpp-1.1/src/PThreads.cpp
@@ -7,7 +7,7 @@ namespace log4cpp {
 
         std::string getThreadId() {
             char buffer[16];
-            ::sprintf(buffer, "%lu", pthread_self());	// thread id unsigned
+            ::sprintf(buffer, "%zd", (size_t) pthread_self());	// thread id unsigned
             return std::string(buffer);     
         }
